base: move specialization type trait to base
