+++ src/v/reflection/type_traits.h
@@ -12,6 +12,7 @@
 #pragma once
 
 #include "base/seastarx.h"
+#include "base/type_traits.h"
 #include "container/fragmented_vector.h"
 #include "tristate.h"
 #include "utils/named_type.h"
@@ -28,14 +29,6 @@
 #include <vector>
 
 namespace detail {
-
-template<typename T, template<typename...> class C>
-struct is_specialization_of : std::false_type {};
-template<template<typename...> class C, typename... Args>
-struct is_specialization_of<C<Args...>, C> : std::true_type {};
-template<typename T, template<typename...> class C>
-inline constexpr bool is_specialization_of_v
-  = is_specialization_of<T, C>::value;
 
 template<class T, template<class, size_t> class C>
 struct is_specialization_of_sized : std::false_type {};
